requirements.txt
scikit-learn
xgboost
PyYAML
jupyter
ipykernel
boto3
Flask
Flask_Cors
joblib
lightgbm
matplotlib
mlflow
dvc
dvc[s3]
numpy
pandas
seaborn